lambdas/event_ingest/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
jsonschema>=4.0.0
kafka-python>=2.0.2
aws-msk-iam-sasl-signer-python>=1.0.0
